feat(swap-widget): add external wallet demo with shared customizer
Adds a second demo route (#external) that exercises the widget's external
wallet path — the demo page itself owns Reown AppKit and feeds walletClient
into <SwapWidget>. Both demos share a single theme/partner customizer panel
extracted to DemoCustomizer.
